corrected checkstyles
