+++ app/build.gradle
@@ -8,7 +8,7 @@ android {
     compileSdk 31
 
     defaultConfig {
-        applicationId "com.example.unser_hoersaal"
+        applicationId "com.example.unserhoersaal"
         minSdk 29
         targetSdk 30
         versionCode 1
