+++ app/src/main/java/com/example/unserhoersaal/model/DatabaseCurrentCourse.java
@@ -1,33 +1,27 @@
-package com.example.unser_hoersaal.model;
+package com.example.unserhoersaal.model;
 
-import android.app.Application;
+
 import android.util.Log;
 
 import androidx.lifecycle.MutableLiveData;
 
-import com.google.firebase.database.ChildEventListener;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
 
-import java.text.SimpleDateFormat;
 import java.util.ArrayList;
-import java.util.Date;
-import java.util.HashMap;
-import java.util.List;
 
-public class DBCurrentCourse {
+public class DatabaseCurrentCourse {
     private FirebaseDatabase firebaseDB;
     private DatabaseReference databaseReference;
     private final MutableLiveData<ArrayList> messages = new MutableLiveData<ArrayList>();
     private ArrayList<Message> messagesList = new ArrayList<Message>();
-    String courseId;
+    private String courseId;
 
 
-
-    public DBCurrentCourse(String courseId) {
+    public DatabaseCurrentCourse(String courseId) {
         this.firebaseDB = FirebaseDatabase.getInstance("https://unser-horsaal-default-rtdb.europe-west1.firebasedatabase.app");
         this.databaseReference = firebaseDB.getReference();
         this.courseId = courseId;
@@ -36,7 +30,7 @@ public class DBCurrentCourse {
             @Override
             public void onDataChange(DataSnapshot dataSnapshot) {
                 messagesList.clear();
-                for (DataSnapshot messageSnapshot: dataSnapshot.getChildren()) {
+                for (DataSnapshot messageSnapshot : dataSnapshot.getChildren()) {
                     messagesList.add(messageSnapshot.getValue(Message.class));
                 }
                 messages.setValue(messagesList);
@@ -51,13 +45,13 @@ public class DBCurrentCourse {
         databaseReference.child("Courses").child(courseId).child("Messages").addValueEventListener(messageListener);
     }
 
-    public void sendMessage(String messageText){
+    public void sendMessage(String messageText) {
         Long time = System.currentTimeMillis();
         Message message = new Message(messageText, time);
         databaseReference.child("Courses").child(courseId).child("Messages").push().setValue(message);
     }
 
-    public MutableLiveData<ArrayList> getMessages(){
+    public MutableLiveData<ArrayList> getMessages() {
         return messages;
     }
