+++ app/src/main/java/com/example/unserhoersaal/viewmodel/CreateCourseViewModel.java
@@ -1,0 +1,29 @@
+package com.example.unserhoersaal.viewmodel;
+
+import androidx.lifecycle.MutableLiveData;
+import androidx.lifecycle.ViewModel;
+
+import com.example.unserhoersaal.model.DatabaseCourseCreation;
+
+
+public class CreateCourseViewModel extends ViewModel {
+    private DatabaseCourseCreation databaseCourseCreation;
+    private final MutableLiveData<String> courseId = new MutableLiveData<String>();
+
+
+    public CreateCourseViewModel() {
+        databaseCourseCreation = new DatabaseCourseCreation();
+    }
+
+    public void createCourse(String courseName, String courseDescription) {
+        databaseCourseCreation.createNewCourse(courseName, courseDescription);
+    }
+
+    public String getCourseId(){
+        return databaseCourseCreation.getCourseId();
+    }
+
+    public void setCourseId(String courseId){
+        databaseCourseCreation.setCourseId(courseId);
+    }
+}
