+++ app/src/main/java/com/example/unserhoersaal/viewmodel/CurrentCourseVMFactory.java
@@ -1,4 +1,4 @@
-package com.example.unser_hoersaal.viewmodel;
+package com.example.unserhoersaal.viewmodel;
 
 import androidx.lifecycle.ViewModel;
 import androidx.lifecycle.ViewModelProvider;
