+++ app/src/main/java/com/example/unserhoersaal/viewmodel/CurrentCourseViewModel.java
@@ -1,21 +1,20 @@
-package com.example.unser_hoersaal.viewmodel;
+package com.example.unserhoersaal.viewmodel;
 
 
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
-import com.example.unser_hoersaal.model.DBCurrentCourse;
+import com.example.unserhoersaal.model.DatabaseCurrentCourse;
 
 import java.util.ArrayList;
-import java.util.List;
 
 
 public class CurrentCourseViewModel extends ViewModel {
-    private DBCurrentCourse databaseCurrentCourse;
+    private DatabaseCurrentCourse databaseCurrentCourse;
     String courseId;
 
     public CurrentCourseViewModel(String courseId) {
         this.courseId = courseId;
-        databaseCurrentCourse = new DBCurrentCourse(courseId);
+        databaseCurrentCourse = new DatabaseCurrentCourse(courseId);
 
     }
