+++ app/src/main/java/com/example/unserhoersaal/viewmodel/EnterCourseViewModel.java
@@ -1,0 +1,21 @@
+package com.example.unserhoersaal.viewmodel;
+
+
+import androidx.lifecycle.ViewModel;
+
+import com.example.unserhoersaal.model.DatabaseCourseCreation;
+
+
+public class EnterCourseViewModel extends ViewModel {
+    private DatabaseCourseCreation databaseCourseCreation;
+    String courseId;
+
+    public EnterCourseViewModel() {
+    }
+
+
+    public void sendCourseId(String courseId){
+
+    }
+
+}
