+++ app/src/main/java/com/example/unserhoersaal/viewmodel/LoginRegisterViewModel.java
@@ -1,4 +1,4 @@
-package com.example.unser_hoersaal.viewmodel;
+package com.example.unserhoersaal.viewmodel;
 
 import android.app.Application;
 
@@ -6,7 +6,7 @@ import androidx.annotation.NonNull;
 import androidx.lifecycle.AndroidViewModel;
 import androidx.lifecycle.MutableLiveData;
 
-import com.example.unser_hoersaal.model.AuthAppRepository;
+import com.example.unserhoersaal.model.AuthAppRepository;
 import com.google.firebase.auth.FirebaseUser;
 
 //source: https://github.com/learntodroid/FirebaseAuthLoginRegisterMVVM/blob/master/app/src/main/java/com/learntodroid/firebaseauthloginregistermvvm/view/LoginRegisterFragment.java [30.12.2021]
