+++ app/src/main/java/com/example/unserhoersaal/views/LoginFragment.java
@@ -1,4 +1,4 @@
-package com.example.unser_hoersaal.views;
+package com.example.unserhoersaal.views;
 
 import android.os.Bundle;
 
@@ -21,16 +21,14 @@ import android.widget.Toast;
 
 
 
-import com.example.unser_hoersaal.R;
-import com.example.unser_hoersaal.viewmodel.LoggedInViewModel;
-import com.example.unser_hoersaal.viewmodel.LoginRegisterViewModel;
+import com.example.unserhoersaal.R;
+import com.example.unserhoersaal.viewmodel.LoggedInViewModel;
+import com.example.unserhoersaal.viewmodel.LoginRegisterViewModel;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.database.DataSnapshot;
-import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
-import com.google.firebase.database.ValueEventListener;
 
 
 public class LoginFragment extends Fragment {
