+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/loginFragment"
-        android:name="com.example.unser_hoersaal.views.LoginFragment"
+        android:name="com.example.unserhoersaal.views.LoginFragment"
         android:label="fragment_login"
         tools:layout="@layout/fragment_login" >
         <action
@@ -19,7 +19,7 @@
     </fragment>
     <fragment
         android:id="@+id/registrationFragment"
-        android:name="com.example.unser_hoersaal.views.RegistrationFragment"
+        android:name="com.example.unserhoersaal.views.RegistrationFragment"
         android:label="fragment_registration"
         tools:layout="@layout/fragment_registration" >
         <action
@@ -31,7 +31,7 @@
     </fragment>
     <fragment
         android:id="@+id/enterOrCreateCourseFragment"
-        android:name="com.example.unser_hoersaal.views.EnterOrCreateCourseFragment"
+        android:name="com.example.unserhoersaal.views.EnterOrCreateCourseFragment"
         android:label="fragment_enter_or_create_course"
         tools:layout="@layout/fragment_enter_or_create_course" >
         <action
@@ -43,7 +43,7 @@
     </fragment>
     <fragment
         android:id="@+id/enterCourseFragment"
-        android:name="com.example.unser_hoersaal.views.EnterCourseFragment"
+        android:name="com.example.unserhoersaal.views.EnterCourseFragment"
         android:label="fragment_enter_course"
         tools:layout="@layout/fragment_enter_course" >
         <action
@@ -52,7 +52,7 @@
     </fragment>
     <fragment
         android:id="@+id/createCourseFragment"
-        android:name="com.example.unser_hoersaal.views.CreateCourseFragment"
+        android:name="com.example.unserhoersaal.views.CreateCourseFragment"
         android:label="fragment_create_course"
         tools:layout="@layout/fragment_create_course" >
         <action
@@ -61,7 +61,7 @@
     </fragment>
     <fragment
         android:id="@+id/currentCourseFragment"
-        android:name="com.example.unser_hoersaal.views.CurrentCourseFragment"
+        android:name="com.example.unserhoersaal.views.CurrentCourseFragment"
         android:label="fragment_current_course"
         tools:layout="@layout/fragment_current_course" />
 </navigation>
